requirements.txt
python-telegram-bot[job-queue]==20.7
sqlalchemy==2.0.25
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-cloud-speech==2.24.1
google-cloud-texttospeech==2.15.1
alembic==1.13.1
python-levenshtein==0.23.0